refactor: have historical sync use indexer funcs
