+++ packages/api/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "api",
+  "name": "@pokt.metrics/api",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
