--- packages/historical-sync/.env.sample
@@ -1,8 +1,0 @@
-ETHEREUM_RPC_URL=
-THE_GRAPH_API_KEY=
-DATABASE_SSL_ENABLED=false
-DATABASE_NAME=defi_dashboard
-DATABASE_PORT=5432
-DATABASE_HOST=db
-DATABASE_PASSWORD=supersecret
-DATABASE_USER=postgres
