+++ packages/historical-sync/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "historical-sync",
+  "name": "@pokt.metrics/historical-sync",
   "version": "1.0.0",
   "description": "",
   "scripts": {
@@ -19,6 +19,7 @@
     "typescript": "^5.8.3"
   },
   "dependencies": {
+    "@pokt.metrics/indexer": "workspace:*",
     "dotenv": "^16.5.0",
     "graphql": "^16.11.0",
     "graphql-request": "^7.2.0",
