--- packages/historical-sync/src/fetchers/index.ts
@@ -1,1 +1,0 @@
-export * from './fetchPoolSnapshot';
