--- packages/historical-sync/src/services/index.ts
@@ -1,1 +1,0 @@
-export * from './storePoolSnapshots';
